cursortut/requirements.txt
# 井字棋游戏依赖包
# 基础依赖
tkinter  # Python内置，无需安装

# 图像处理
Pillow>=9.0.0

# 可选依赖（用于增强功能）
numpy>=1.21.0  # 用于高级图像处理
opencv-python>=4.5.0  # 用于图像特效（可选）

# 开发依赖（可选）
pytest>=6.0.0  # 用于测试
black>=21.0.0  # 代码格式化
flake8>=3.9.0  # 代码检查
